Updated stg_erp__salesorderdetail.sql and int_sales__salesmetrics.sql…

+++ models/staging/erp/stg_erp__salesorderdetail.sql
@@ -13,7 +13,7 @@ renamed as (
         cast(SALESORDERDETAILID as int) as salesorderdetailid_pk
         , cast(SALESORDERID as int) as salesorderid_fk
         , cast(ORDERQTY as int) as order_quantity
-        , cast(UNITPRICE as int) as unitprice
+        , cast(UNITPRICE as float) as unitprice
         , cast(PRODUCTID as int) as productid
         , cast(UNITPRICEDISCOUNT as int) as unitpricediscount
